Refactorización de Builders

+++ Fire-Emblem/Effects/Damage/ExtraDamage/BackAtYouEffect.cs
@@ -14,7 +14,7 @@ public class BackAtYouEffect : IExtraDamageEffect
     public void ApplyEffect(Unit activator, Unit opponent)
     {
         Unit targetUnit = Target == EffectTarget.Unit ? activator : opponent;
-        int extraDamage = (int)((activator.BaseHp - activator.CurrentHP) / 2);
+        int extraDamage = (activator.BaseHp - activator.CurrentHP) / 2;
         targetUnit.ApplyExtraDamageEffect(extraDamage);
         targetUnit.AddActiveEffect(this);
     }
